IGNITE-24847 Move CF to org.apache.ignite.internal.util.typedef
